ImageProsessing/p_ImageRestoration/ILL_Posed/Deblur_DL/Self-Supervised Blind Motion Deblurring With Deep Expectation Maximization/requirements.txt
opencv-python==4.2.0.32
scikit-image==0.17.2
scipy==1.4.1
numpy==1.19.5
natsort==7.1.0
tqdm==4.43.0
tensorboardX==2.5.1
torchvision==0.8.2



